Coupled "PricingModel" and "Market"
