+++ src/elfpy/policies/no_action.py
@@ -6,7 +6,6 @@ subclasses of BasicPolicy will implement trade actions
 """
 
 from elfpy.markets import Market, MarketAction
-from elfpy.pricing_models.base import PricingModel
 from elfpy.agent import Agent
 
 
@@ -15,7 +14,7 @@ class NoAction(Agent):
     Most basic policy setup, which implements a noop agent that performs no action
     """
 
-    def action(self, market: Market, pricing_model: PricingModel) -> list[MarketAction]:
+    def action(self, market: Market) -> list[MarketAction]:
         """Returns an empty list, indicating now action"""
         # pylint disable=unused-argument
         action_list = []
