+++ src/elfpy/policies/single_short.py
@@ -6,7 +6,6 @@ User strategy that opens a single short and doesn't close until liquidation
 
 from elfpy.agent import Agent
 from elfpy.markets import Market
-from elfpy.pricing_models.base import PricingModel
 from elfpy.types import MarketActionType
 
 
@@ -18,7 +17,7 @@ class Policy(Agent):
         self.pt_to_short = 100
         super().__init__(wallet_address, budget)
 
-    def action(self, market: Market, pricing_model: PricingModel):
+    def action(self, market: Market):
         """
         implement user strategy
         short if you can, only once
@@ -26,7 +25,7 @@ class Policy(Agent):
         action_list = []
         block_position_list = list(self.wallet.shorts.values())
         has_opened_short = bool(any((x < -1 for x in block_position_list)))
-        can_open_short = self.get_max_pt_short(market, pricing_model) >= self.pt_to_short
+        can_open_short = self.get_max_pt_short(market) >= self.pt_to_short
         if can_open_short and not has_opened_short:
             action_list.append(
                 self.create_agent_action(action_type=MarketActionType.OPEN_SHORT, trade_amount=self.pt_to_short)
